Addressing small formatting feedback

+++ src/c++/perf_analyzer/constants.h
@@ -1,4 +1,4 @@
-// Copyright 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2022-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -38,7 +38,7 @@ constexpr static const uint32_t OPTION_ERROR = 3;
 
 constexpr static const uint32_t GENERIC_ERROR = 99;
 
-const float DELAY_PCT_THRESHOLD{1.0};
+const double DELAY_PCT_THRESHOLD{1.0};
 
 /// Different measurement modes possible.
 enum MeasurementMode { TIME_WINDOWS = 0, COUNT_WINDOWS = 1 };
